[feat] helm-model.ts 추가
